(PC-20937)[PRO] refactor: move config to TS
