+++ pro/.storybook/main.ts
@@ -1,0 +1,19 @@
+import type { StorybookConfig } from '@storybook/react-webpack5'
+
+const config: StorybookConfig = {
+  stories: [
+    '../src/ui-kit/Icons/Icons.stories.tsx',
+    '../src/**/*.stories.@(js|jsx|ts|tsx)',
+  ],
+  addons: ['@storybook/addon-a11y', '@storybook/preset-scss'],
+  staticDirs: ['../public'],
+  framework: {
+    name: '@storybook/react-webpack5',
+    options: {},
+  },
+  docs: {
+    autodocs: true,
+  },
+}
+
+export default config
