+++ pro/.storybook/preview-body.html
@@ -1,6 +1,6 @@
 <style>
-  #root {
-    display: block!important;
+  #storybook-root {
+    display: block !important;
     max-width: 960px;
-}
+  }
 </style>
